Use Node in NodeAuthoringController.
Set node and projectId in resolve of the UI router paths so they
are available in the controller as bindings.
Kept nodeJson in nodeAuthoringController until we have a good
solution for updating node in cache.
Defined Node.isGroup() and updated references.
Removed unused functions and dependencies.
Updated path from '/wise5' to '/assets/wise5'

#102