build(NPM): Run npm install inside Docker container (#1024)
Upgrading to Node 18 created a new issue with hot-reload during development where page did not completely render. This is likely due to the new node version. The customElement re-definition bug is an on-going bug that can be worked-around for now by refreshing the page if one needs to test the customElements (milestone data and graph components)